Add ch03/ex09 code.

+++ ch03/mandelbrot/mandelbrotEx08.go
@@ -21,8 +21,8 @@ var FormatCompTest FormatType = DEFAULT
 var IterationsVal = 200
 
 func compCmplxFormat(px, py int) color.Color {
-	y := float64(py)/float64(Height)*(YMax-YMin) + YMin
-	x := float64(px)/float64(Width)*(XMax-XMin) + XMin
+	y := float64(py)/Height*(YMax-YMin) + YMin
+	x := float64(px)/Width*(XMax-XMin) + XMin
 
 	var color color.Color
